pyproject.toml
[tool.poetry]
name = "certification-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.10"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
django-filter = "^24.2"
pytest = "^8.2.0"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
